requirements.txt
websocket-client>=0.57.0,<2
base58>=1.0.3,<3
certifi>=2019.3.9
idna>=2.1.0,<4
requests>=2.21.0,<3
xxhash>=1.3.0,<4
pytest>=4.4.0
ecdsa>=0.17.0,<1
eth-keys>=0.2.1,<1
eth_utils>=1.3.0,<3
pycryptodome>=3.11.0,<4
PyNaCl>=1.0.1,<2
substrate-interface==1.7.5

scalecodec>=1.2.8,<1.3
py-sr25519-bindings>=0.2.0,<1
py-ed25519-zebra-bindings>=1.0,<2
py-bip39-bindings>=0.1.9,<1

mkdocs
mkdocs-material
mkdocs-autorefs
mkdocstrings
mkdocstrings[python]
